+++ src/r0/directory.rs
@@ -2,6 +2,8 @@
 
 pub mod get_public_rooms;
 pub mod get_public_rooms_filtered;
+pub mod get_room_visibility;
+pub mod set_room_visibility;
 
 use js_int::UInt;
 use ruma_identifiers::{RoomAliasId, RoomId};